[ISSUE#5696] Move and repackage some ram class to ram package. (#7040)
* move client.identity package to auth.ram.identity.

* move SignUtil to ram.utils.

* Add shutdown for auth plugin.

* Move SpasAdapter to new package.
