+++ client/src/main/java/com/alibaba/nacos/client/config/http/ServerHttpAgent.java
@@ -19,7 +19,6 @@ package com.alibaba.nacos.client.config.http;
 import com.alibaba.nacos.api.exception.NacosException;
 import com.alibaba.nacos.client.config.impl.ConfigHttpClientManager;
 import com.alibaba.nacos.client.config.impl.ServerListManager;
-import com.alibaba.nacos.client.config.impl.SpasAdapter;
 import com.alibaba.nacos.client.utils.ContextPathUtil;
 import com.alibaba.nacos.client.utils.LogUtils;
 import com.alibaba.nacos.client.utils.ParamUtil;
@@ -48,10 +47,6 @@ public class ServerHttpAgent implements HttpAgent {
     
     private static final NacosRestTemplate NACOS_RESTTEMPLATE = ConfigHttpClientManager.getInstance()
             .getNacosRestTemplate();
-    
-    private String accessKey;
-    
-    private String secretKey;
     
     private String encode;
     
@@ -279,7 +274,6 @@ public class ServerHttpAgent implements HttpAgent {
         String className = this.getClass().getName();
         LOGGER.info("{} do shutdown begin", className);
         ConfigHttpClientManager.getInstance().shutdown();
-        SpasAdapter.freeCredentialInstance();
         LOGGER.info("{} do shutdown stop", className);
     }
 }
