+++ client/src/test/java/com/alibaba/nacos/client/auth/ram/injector/NamingResourceInjectorTest.java
@@ -20,7 +20,7 @@ import com.alibaba.nacos.api.PropertyKeyConst;
 import com.alibaba.nacos.client.auth.LoginIdentityContext;
 import com.alibaba.nacos.client.auth.ram.RamContext;
 import com.alibaba.nacos.client.auth.spi.RequestResource;
-import com.alibaba.nacos.client.naming.utils.SignUtil;
+import com.alibaba.nacos.client.auth.ram.utils.SignUtil;
 import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
