+++ client/src/test/java/com/alibaba/nacos/client/naming/remote/AbstractNamingClientProxyTest.java
@@ -27,7 +27,7 @@ import com.alibaba.nacos.api.naming.pojo.ServiceInfo;
 import com.alibaba.nacos.api.selector.AbstractSelector;
 import com.alibaba.nacos.client.auth.spi.RequestResource;
 import com.alibaba.nacos.client.naming.event.ServerListChangedEvent;
-import com.alibaba.nacos.client.naming.utils.SignUtil;
+import com.alibaba.nacos.client.auth.ram.utils.SignUtil;
 import com.alibaba.nacos.client.security.SecurityProxy;
 import com.alibaba.nacos.client.utils.AppNameUtils;
 import com.alibaba.nacos.common.notify.Event;
